requirements.txt
asgiref==3.7.2
astroid==3.0.2
certifi==2023.7.22
chardet==3.0.4
colorama==0.4.6
dill==0.3.7
distlib==0.3.7
Django==4.2.7
djangorestframework==3.14.0
django-extensions==3.1.3
filelock==3.12.4
gunicorn==21.2.0
idna==2.7
iniconfig==2.0.0
isort==5.13.1
marvel==0.1.0
maturin==1.4.0
mccabe==0.7.0
mysqlclient==2.1.0
packaging==23.2
pillow==10.2.0
platformdirs==3.11.0
pluggy==1.3.0
psutil==5.9.8
psycopg2-binary==2.9.9
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyperf==2.6.2
pytest==7.4.4
python-dotenv==0.19.2
pytz==2021.3
requests==2.20.0
sqlparse==0.4.2
tomlkit==0.12.3
tzdata==2023.3
urllib3==1.24.3
virtualenv==20.24.5




